Added Alert Dialog to grocery delete

+++ app/src/main/java/com/example/capstoneproject/GroceryActivity.java
@@ -6,6 +6,8 @@ import androidx.appcompat.widget.Toolbar;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Parcelable;
@@ -13,6 +15,7 @@ import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.ListView;
+import android.widget.Toast;
 
 import com.example.capstoneproject.Models.RandomRecipe;
 import com.example.capstoneproject.Models.RecipeNutritionResponse;
@@ -101,9 +104,18 @@ public class GroceryActivity extends AppCompatActivity {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
                 if (item.getItemId() == R.id.grocery_clear) {
-                    groceryDao.deleteAll();
-                    groceries.clear();
-                    adapter.notifyDataSetChanged();
+                    AlertDialog.Builder builder = new AlertDialog.Builder(adapter.getContext());
+                    builder.setTitle("Delete Recipe");
+                    builder.setMessage("Are you sure you want to clear the grocery list?");
+                    builder.setPositiveButton("Confirm", (DialogInterface dialogInterface, int i) -> {
+                        groceryDao.deleteAll();
+                        groceries.clear();
+                        adapter.notifyDataSetChanged();
+                    });
+                    builder.setNegativeButton(android.R.string.cancel, (DialogInterface dialogInterface, int i) ->
+                            adapter.notifyDataSetChanged());
+                    AlertDialog dialog = builder.create();
+                    dialog.show();
                     return true;
                 }
                 return false;
